get all users is working :)
